requirements-dev.txt
# Zależności dla lokalnego developmentu (Windows/Mac/Linux)
Django==4.2.7
Pillow==11.0.0
python-decouple==3.8

